requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.15.0
    # via
    #   tensorboard
    #   tensorflow
astor==0.8.1
    # via tensorflow
attrs==21.4.0
    # via jsonschema
cached-property==1.5.2
    # via h5py
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
colorama==0.4.4
    # via tqdm
dataclasses==0.8
    # via werkzeug
gast==0.2.2
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.44.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.1.0
    # via keras-applications
idna==3.3
    # via requests
importlib-metadata==4.8.3
    # via
    #   jsonschema
    #   markdown
importlib-resources==5.4.0
    # via tqdm
joblib==1.1.0
    # via scikit-learn
jsonschema==3.2.0
    # via valohai-yaml
keras-applications==1.0.8
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.6
    # via tensorboard
numpy==1.19.5
    # via
    #   -r requirements.in
    #   h5py
    #   keras-applications
    #   keras-preprocessing
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
opt-einsum==3.3.0
    # via tensorflow
pandas==1.1.5
    # via -r requirements.in
pillow==8.4.0
    # via -r requirements.in
protobuf==3.19.4
    # via
    #   tensorboard
    #   tensorflow
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via valohai-yaml
requests==2.27.1
    # via valohai-utils
scikit-learn==0.24.2
    # via sklearn
scipy==1.5.4
    # via scikit-learn
six==1.16.0
    # via
    #   absl-py
    #   google-pasta
    #   grpcio
    #   jsonschema
    #   keras-preprocessing
    #   python-dateutil
    #   tensorboard
    #   tensorflow
sklearn==0.0
    # via -r requirements.in
tensorboard==1.15.0
    # via tensorflow
tensorflow==1.15.0
    # via -r requirements.in
tensorflow-estimator==1.15.1
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
threadpoolctl==3.1.0
    # via scikit-learn
tqdm==4.64.0
    # via valohai-utils
typing-extensions==4.1.1
    # via importlib-metadata
urllib3==1.26.9
    # via requests
valohai-utils==0.1.11
    # via -r requirements.in
valohai-yaml==0.15.0
    # via valohai-utils
werkzeug==2.0.3
    # via tensorboard
wheel==0.37.1
    # via
    #   tensorboard
    #   tensorflow
wrapt==1.14.0
    # via tensorflow
zipp==3.6.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
